Glamour Salon/requirements.txt
streamlit
pandas
numpy
plotly
Pillow
setuptools
wheel